Move tests into news directory structure
